Fix to find the correct duration of ogg files with the last granule position.
While finding the duration of ogg files, the last granule position is searched. While
doing this, finding EOF is considered as an error and it is returned. This causes the
loss of audio for app 'Three little Shelf drum'. Whereas EOF should be considered as
a warning and the granule position should be searched again by moving the current pointer
still back. This fixes the issue.

Change-Id: Ic90daa971c181fccdbf3260fa7ce41935a319673
Signed-off-by: Muthukumar K <muthukumar.kandasamy@intel.com>
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>
Signed-off-by: Yong Yao <yong.yao@intel.com>