Fix bugs regarding restoreRecord_l(start)
AudioRecord mute issue when recorded device is changed.

After finishing recording, if a user changes the recorded device to another, and tries to record again, this issue has occured(time sensitive issue)
The mCblk->mFlags is CBLK_INVALID value, and restoreRecord_l(start) is called in AudioRecord::start() api.
openRecord_l() is successfully called, but the mAudioRecord->start() is not called becuase the mActive value if false.
The return value of restoreRecord_l(start) api is NO_ERROR, but audio is not recorded(bacause the mAudioRecord->start() is not called)

Change-Id: I0da9a02b0148ff340f40cc6166774f8b6140459e
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>